refactor: 결과 확인 후 당첨 개수 및 금액 반환 함수 분리
