+++ src/controllers/lottoController.js
@@ -1,6 +1,5 @@
-import checkRank from "../domain/checkRank.js";
+import checkWinningResults from "../domain/checkWinningResults.js";
 import { generateLottos } from "../domain/generateLottos.js";
-import updateMatchResult from "../domain/updateMatchResult.js";
 import { getBonusNumber, getWinningNumbers } from "../view/input.js";
 import {
   printLottoCount,
@@ -17,22 +16,9 @@ const lottoController = async (price) => {
   const winningNumbers = await getWinningNumbers();
   const bonusNumber = await getBonusNumber(winningNumbers);
 
-  lottos.forEach((lotto) => {
-    updateMatchResult(lotto, winningNumbers, bonusNumber);
-  });
-
-  let totalReward = 0;
-  let resultCount = new Array(6).fill(0);
-
-  lottos.forEach((lotto) => {
-    const result = checkRank(
-      lotto.matchResult.matchCount,
-      lotto.matchResult.isBonusMatched
-    );
-    if (result) {
-      totalReward += result.REWARD;
-      resultCount[result.RANK]++;
-    }
+  const { resultCount, totalReward } = checkWinningResults(lottos, {
+    winningNumbers,
+    bonusNumber,
   });
 
   printResult(resultCount);
